Move entities from feature-task to core-entities
